worker max answers now works
Now every user can complete the a task only n times, where ne is defined
when the task is created.
There still is a problem because the control is done in the tasklist and
after finishing a task you can continue (without returning to the
tasklist). Should be easy to fix will do asap.